--- a/src/alma/benchmark/log.py
+++ b/src/alma/benchmark/log.py
@@ -64,9 +64,6 @@ def log_results(
     display_function(f"Device: {results['device']}")
     display_function(f"Total elapsed time: {results['total_elapsed_time']:.3f} seconds")
     display_function(
-        f"Mean batch time: {results['mean_batch_time']:.3f} seconds ± {results['batch_time_std']:.3f}"
-    )
-    display_function(
         f"Total samples: {results['total_samples']} - Batch size: {results['batch_size']}"
     )
     display_function(f"Data dtype: {results['data_dtype']}")