## Minor example updates
Updated some examples, e.g. added multiple workers and pinned memory details to data loader example.

## More accurate throughput measurements

The old code only accumulates the time spent in forward passes, which is artificially fast.

The new code includes:
  - Data transfer time (`.to(device)`)
  - CUDA queue scheduling overhead
  - Any inter-batch synchronization/overhead
  - DataLoader iteration overhead


**Previous timings example:**
```
COMPILE_INDUCTOR_MAX_AUTOTUNE results:
Device: cuda
Total elapsed time: 0.012508 seconds
Total inference time (model only): 0.003409 seconds
Total samples: 2048 - Batch size: 64
Data dtype: torch.float32
Throughput: 600846.05 samples/second


COMPILE_OPENXLA results:
Device: xla:0
Total elapsed time: 0.013253 seconds
Total inference time (model only): 0.003098 seconds
Total samples: 2048 - Batch size: 64
Data dtype: torch.float32
Throughput: 661091.90 samples/second

```

**New timings:**
``` 
COMPILE_INDUCTOR_MAX_AUTOTUNE results:
Device: cuda
Total elapsed time: 0.012 seconds
Total samples: 2048 - Batch size: 64
Data dtype: torch.float32
Throughput: 174216.03 samples/second


COMPILE_OPENXLA results:
Device: xla:0
Total elapsed time: 0.011 seconds
Total samples: 2048 - Batch size: 64
Data dtype: torch.float32
Throughput: 179965.56 samples/second
```